backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
pandas==2.1.3
numpy==1.24.3
pydantic==2.5.0
python-multipart==0.0.6